* Updated bet price to 3 coins.  * Converted create function in betService to purchaseBet to run transaction of removing coins from users account.  * Added initialCoinStart amount.  * Started BetPurchaseSuccessState to display after a user places a random bet.  * Converted betPrice on GameModel from double to int.  * Added coins to UserModel.

+++ lib/models/game_model.dart
@@ -6,7 +6,7 @@ class GameModel {
   int awayTeamScore;
   int homeTeamID;
   int homeTeamScore;
-  double betPrice;
+  int betPrice;
   int status;
   int betCount;
